# soya-framework-admin/src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "title" : "Workshop Server",
    "contact" : {
      "name" : "wenqun.soya@gmail.com"
    }
  },
  "basePath" : "/api",
...
  "schemes" : [ "http" ],
  "paths" : {
...
    "/cmd/execute/{group}/{command}" : {
      "post" : {
        "tags" : [ "Command Service" ],
        "operationId" : "execute",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "group",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "command",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "arguments",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/cmd/uri" : {
      "post" : {
        "tags" : [ "Command Service" ],
        "operationId" : "uri",
        "consumes" : [ "text/plain" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/cmd/command/{group}/{command}" : {
      "get" : {
        "tags" : [ "Command Service" ],
        "operationId" : "command",
        "consumes" : [ "text/plain" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "group",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "command",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/cmd/commands" : {
      "get" : {
        "tags" : [ "Command Service" ],
        "operationId" : "commands",
        "consumes" : [ "text/plain" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "group",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/kafka/poll-to-end/{environment}" : {
      "get" : {
        "tags" : [ "kafka" ],
        "operationId" : "pollToEnd",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "environment",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "timeout",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/kafka/pub-and-sub/{environment}" : {
      "post" : {
        "tags" : [ "kafka" ],
        "operationId" : "pubAndSub",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "environment",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "timeout",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "produceTopic",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "consumeTopic",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/kafka/topics/{environment}" : {
      "get" : {
        "tags" : [ "kafka" ],
        "operationId" : "topics",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "environment",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "timeout",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "query",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/kafka/topic/{environment}" : {
      "get" : {
        "tags" : [ "kafka" ],
        "operationId" : "topic",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "environment",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "timeout",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "topicName",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/kafka/consume/{environment}/{consumeTopic}" : {
      "get" : {
        "tags" : [ "kafka" ],
        "operationId" : "consume",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "environment",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "consumeTopic",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "timeout",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "format",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/kafka/produce/{environment}" : {
      "post" : {
        "tags" : [ "kafka" ],
        "operationId" : "produce",
        "consumes" : [ "text/plain", "application/json", "application/xml" ],
        "produces" : [ "text/plain", "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "environment",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "timeout",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "produceTopic",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
  }
}